+++ example/lib/screens/login_screen.dart
@@ -1,7 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:example/router.dart';
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 class LoginScreen extends StatelessWidget {
   final double id;
@@ -24,11 +23,6 @@ class LoginScreen extends StatelessWidget {
           child: FlatButton(
             child: Text("Login"),
             onPressed: () async {
-              final prefs = await SharedPreferences.getInstance();
-              prefs.setString('token', 'token value');
-              // ExtendedNavigator.of(context)
-              //     .pushReplacementNamed(Routes.homeScreen);
-              // ExtendedNavigator.of(context).pop(true);
               ExtendedNavigator.of(context)
                   .pushReplacementNamed(Routes.homeScreen);
             },