+++ example/lib/screens/nested_screens/nested_router.dart
@@ -3,7 +3,7 @@ import 'package:auto_route/auto_route_annotations.dart';
 import 'nested_screen.dart';
 import 'nested_screen_two.dart';
 
-@MaterialAutoRouter()
+@MaterialAutoRouter(generateNavigationHelperExtension: true)
 class $NestedRouter {
   @initial
   NestedScreen nestedScreen;